bash prompt: get branch name via a new 'git prompt--helper' builtin command
Add the 'git prompt--helper' builtin command to eventually replace most of
the shell logic with all its fork()+exec()s with a single command.

So far it only prints the branch name, but thus it already replaces the
HEAD parsing with bash builtins and the 'git symbolic-ref' execution for
symlinked symrefs.

Signed-off-by: SZEDER Gábor <szeder@ira.uka.de>