bash prompt: remove superfluous redirection and fix && chain in test
The test 'prompt - bash color pc mode - branch name' redirected the output
of __git_ps1() into a file, only to overwrite it in the very next line.  A
prior test already checks that __git_ps1() doesn't produce any output in
PROMPT_COMMAND mode, so let's just remove the redirection.

The same line was missing the && at the end, fix it.

Signed-off-by: SZEDER Gábor <szeder@ira.uka.de>