Add tests for git-sh-setup's require_clean_work_tree()
Add tests that check require_clean_work_tree() in the common cases,
i.e. on a branch with all combinations of clean and dirty index and
worktree.

Also add failing tests that demonstrate misbehavior on orphan or
unborn branches, i.e. that require_clean_work_tree() always exits with
error on an orphan branch independently from whether the index and
worktree are clean or dirty, and it does so because of the invalid
HEAD, meaning that even when it should exit with error, it does so for
the wrong reason.

Signed-off-by: SZEDER Gábor <szeder@ira.uka.de>