sh-setup: make require_clean_work_tree() work on unborn/orphan branches
git-sh-setup's require_clean_work_tree() always exits with error on an
unborn or orphan branch, even when the index and worktree are both
clean.  The reason is that require_clean_work_tree() starts off with
verifying HEAD, to make sure that it can safely pass HEAD to 'git
diff-index' later when it comes to checking the cleanness of the
index, and errors out finding the invalid HEAD of the orphan branch.
This also means, that even when require_clean_work_tree() should exit
with error, i.e.  when the index or both the index and the worktree
are dirty, it does so for the wrong reason.

There are scripts requiring a clean worktree that should work on an
unborn or orphan branch as well, and those should be able to use this
function instead of duplicating its functionality and nice error
reporting in a way that handles orphan branches.

Fix this by comparing the index to the empty tree while on an unborn
or orphan branch, and to HEAD otherwise.  This makes all three failing
tests added in the previous commit succeed.

Signed-off-by: SZEDER Gábor <szeder@ira.uka.de>