ls-files: warn users that '--error-unmatch' has no effect without pathspecs
If 'git ls-files <other options> --error-unmatch' is executed without
specifying any pathspecs, it lists all paths matching the criteria of
those other options, or if there are no such paths present, then it
exits quietly without an error message or an error code.  IOW
'--error-unmatch' without any pathspecs has no effect whatsoever.

This is most likely not what the user wanted, so add a warning() to
let them know about this behavior.  Furthermore, since this is not
mentioned in the documentation (though arguably it's in there
implicitly), add a sentence about this behavior to document it
explicitly.

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>