coccinelle: exclude sha1dc source files from static analysis
sha1dc is an external library (that we carry in-tree for convenience
or grab as a submodule), so there is no use in applying our semantic
patches to its source files.

Furthermore, presumably because of the heavy use of repetitive macro
declarations, applying the semantic patches 'array.cocci' and
'swap.cocci' to 'sha1dc/sha1.c' takes over half a minute each on my
tiny laptop, which amounts to about a third of the runtime of applying
these two semantic patches to the whole git source tree.

Therefore, exclude sha1dc's source files from Coccinelle's static
analysis.

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>