travis-ci: run tests in shorter paths
In Travis CI build jobs the absolute path to the build directory
includes the GitHub repo slug, i.e. "owner_name/repo_name".  If the
owner and/or repo name is long enough, then the absolute path of a
trash directory can be too long for creating a unix domain socket,
whose name's maximum length is limited to around 104-110 chars,
depending on the platform.  As a results, some tests might be missing
their GPG prereq even if GnuPG is installed, because the 'gpg'
invocations in 'lib-gpg.sh' error out with:

  gpg: can't connect to the agent: File name too long
  gpg: error getting the KEK: No agent running

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>