t/lib-gpg: allow testing with a particular gpg program
GnuPG 2.x has enough differences compared to prior versions that
developers working on GPG-related parts of Git might want to run the
tests with different GPG versions.  While it's certainly already
possible to do so, e.g. by prepending a directory containg the chosen
gpg executable to the $PATH used to run the tests, it's not quite that
convenient.

Let's provide a more convenient way to do so via the
$GIT_TEST_GPG_PROGRAM environment variable, which should be set to the
desired GnuPG executable.  If set, then 'lib-gpg.sh' will use that
executable to determine whether it fulfills the GPG prereq, and will
set it as 'gpg.program' in the config.  This also implies that setting
$GIT_TEST_GPG_PROGRAM to something non-GPG (e.g. 'echo', 'false', or
even something non-existent) will effectively skip any tests depending
on the GPG prereq.

The test 'git tag -s fails if gpg is misconfigured (bad signature
format)' in 't7004-tag.sh' sets 'gpg.program' to an invalid GPG
executable.  It does so using 'test_config', meaning that after that
test 'gpg.program' will be unset.  However, the subsequent test
depends on the GPG prereq, too, so take care to restore the config
variable's previous value if there was one.

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>