treewide: use 'ptr = xmalloc(sizeof(*ptr))'
It's preferable to write single-element memory allocations as:

  ptr = xmalloc(sizeof(*ptr));

instead of

  ptr = xmalloc(sizeof(struct foobar));

because it eliminates the possibility of mistakenly specifying the
wrong data type by inferring it, and because we wouldn't have to touch
the allocation if we were to change 'ptr's data type or rename 'struct
foobar' in the future.

Add a Coccinelle semantic patch 'contrib/coccinelle/alloc.cocci' to
transform the latter to the former, and after reviewing the suggested
transformations apply them to the source tree.

Note that the two transformations in 'compat/mingw.c' and
'compat/precompose_utf8.c' were done manually; I run Coccinelle on
Linux, where those two files are ignored.

Note also that this semantic patch would miss cases where multiple
variables are declared on the same line with the memory allocation,
e.g.:

  int *ptr = xmalloc(sizeof(int)), i, j;

AFAICT there are no such lines in the codebase at the moment.

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>