treewide: use ALLOC_ARRAY even when declaring variables
The semantic patch 'contrib/coccinelle/array.cocci' already contains
two rules to transform certain xmalloc() calls to use the shorter and
safer ALLOC_ARRAY macro.  These rules, however, miss all cases where a
variable's declaration and the memory allocation are on the same line,
e.g. in walker.c:walker_fetch():

  struct object_id *oids = xmalloc(targets * sizeof(struct object_id));

Add two new rules to catch and transform these cases as well, and
after a bit of a touch-up apply the two suggested transformations.

The touch-ups are necessary, because these rules have to split the
variable declaration and the memory allocation to two subsequent
lines, i.e. the above example would be transformed into this:

  struct object_id *oids;
  ALLOC_ARRAY(oids, targets);

Since this is not the last variable declaration in that function, it
would then cause declaration-after-statement errors during
compilation.  As far as static analysis goes that's acceptable,
because the primary goal is to catch undesired code patterns.  And
even if some brave developers were to blindly trust Coccinelle's
suggestions and apply them, the resulting compilation errors will
serve as a loud enough reminder that they should have a closer look.

Note that these new rules would miss cases where multiple variables
are declared on the same line with the memory allocation, e.g.:

  int *ptr = xmalloc(sizeof(int) * size), i, j;

AFAICT there are no such lines in the codebase at the moment.

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>