ci: use Coccinelle v1.0.7 in the static analysis build job
Use Coccinelle v1.0.7, the newest release at the moment, in our static
analysis build job on Travis CI, because it is apparently more
thorough in its analysis and can detect issues that previous versions
overlook (see the previous two patches in this series), and it appears
to be somewhat faster, too.  For reference: we currently use v1.0.0 in
Travis CI's Ubuntu 14.04 images, Ubuntu 16.04 and Debian unstable are
based on v1.0.4.

Alas, I didn't find a package repository from where we could
conveniently 'apt-get install' the newest Coccinelle version [1].  So
I put together a small-ish Docker image 'szeder/coccinelle'
specifically to run 'make coccicheck' with Coccinelle v1.0.7 as its
default command, which can be used in our CI build jobs (or even on
contributors' machines just as well, without polluting their setups
with obscure OCaml dependencies).

Modify the static analysis build job to start the docker service, pull
the aforementioned Docker image in the 'before-install' phase, and run
the container in the 'script' phase.  Take care that the commands
inside the container are not executed as root but as a uid:gid that
matches the host user, so the host user has access to all generated
files and can show Coccinelle's suggested transformations, if any.

[1] Actually, there is a PPA 'ppa:npalix/coccinelle' containing
    packages of the latest Coccinelle version, but only for the latest
    Ubuntu 18.10, so it's far too new to be usable in our CI builds.

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>