t6050-gc: don't check the length of the progress output
The test 'gc --no-quiet' overzealously checks that 'git gc --no-quiet'
produced exactly one line to its standard error, and triggers a
failure in combination with GIT_TEST_GETTEXT_POISON_SCRAMBLED and
commit 545dc345eb (progress: break too long progress bar lines,
2019-04-12): in scrambled mode the progress output becomes longer than
the standard width of a dumb terminal, and thus the progress bar is
split into two lines.

This check is not only useless but misfires, so drop it.

[TODO: should this be squashed into the previous 'gettext: introduce
       GIT_TEST_GETTEXT_POISON_SCRAMBLED' patch?]

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>