SQUASH??? compat/obstack: fix portability issues
This is sort-of a cherry-pick of d190a0875f (obstack: Fix portability
issues, 2011-08-28), which is necessary to make 'compat/obstack.c'
compile again.  Only "sort-of a cherry-pick", because the divergence
between upstream and our copy was just too bit and I gave up on the
conflict resolution, and instead made the still necessary/applicable
edits in the spirit of d190a0875f by hand.

With this patch it can be built with GCC both on Linux and on macOS,
but, alas, not with Clang.

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>