t7030-verify-tag: don't save stdandard error of a loop
In 't7030-verify-tag' the tests 'verify multiple tags' and 'verify
multiple tags x509' save the standard error of a 'for' loop running
'git verify-tag' on different tags to create the tests' expected
results.  Saving the stderr of any compound command is, in general,
asking for trouble, because when the test script is run with '-x'
tracing enabled, then the compound command's stderr includes the trace
of the commands executed as part of the compound command as well,
which caused several issues in the past [1].

Luckily, it doesn't affect these two tests' correctness, because both
'for' loops' stderr is first filtered through a 'grep' command to
extract GNUPG output, throwing away all the undesired '-x' trace.

Still, it sets bad example; and indeed, the test 'verify multiple tags
x509' saves the loop's stderr, because, presumably, it simply followed
suit of 'verify multiple tags', although by the time it was added
't/README' already discouraged saving the stderr of compound commands.

Don't save the stderr of the whole loops, but only the stderr of the
relevant 'git verify-tag' commands in the loops' bodies, and, for
consistency's sake, do the same with their stdout as well.

[1] See a5bf824f3b (t: prevent '-x' tracing from interfering with test
    helpers' stderr, 2018-02-25) and subsequent commits.

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>