ci: extract a function to download and install Perforce
In the Clang and GCC Linux CI jobs we "install" Perforce by
downloading the 'p4d' and 'p4' binaries from a Perforce filehost with
'wget' to a directory in PATH and set their executable bit.  Extract
all this into a helper function, so we can easily do the same in the
OSX CI jobs as well.

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>