travis-ci: build and test on s390x
Travis CI recently started to support multiple CPU architectures [1],
though it's still only in alpha.  Among the new supported
architectures there is one big endian: s390x.  Since we had an issue
or two on big endian archs recently [2], let's make use of it, and
build and test on s390x as well, to get coverage on big endian archs,
and hopefully we'll be able to catch similar issues in a more timely
manner in the future.

And let's hope that being an alpha feature will not cause too many
headaches...

The build and test suite runs surprisingly fast on s390x, in fact it's
the fastest of all our build jobs that run the test suite.  So let's
run the test suite a second time with all the various GIT_TEST_*
environment variables set, like in the 'linux-gcc' job, so those
features get some coverage on a big endian arch as well.

Several tests in 't5319-multi-pack-index.sh' that run 'git
count-objects' with and without multi-pack-index enabled to compare
its output fail rather frequently on s390x (and, for the record, on
arm64 as well, so it's not an endianness issue).  Skip this test
script for now, until we figure out what's going on and fix that
issue.

[1] https://blog.travis-ci.com/2019-10-07-multi-cpu-architecture-support
[2] 2b6f6ea1bd (test-progress: fix test failures on big-endian
    systems, 2019-10-20)
    e1e7a77141 (t: sort output of hashmap iteration, 2019-07-30)

Signed-off-by: SZEDER Gábor <szeder.dev@gmail.com>